+++ docs/CHANGELOG.rst
@@ -1,12 +1,17 @@
 Changelog
 =========
 
-Version 0.4.5
+Version 0.5.0
 -------------
 In development.
 
 * Changes in API:
     - ``Particle.table()`` renamed to ``Particle.all()``.
+* Added the 2019 PDG data table, now default.
+* Enhancements to  ``Particle`` class:
+  - Numerous LaTeX particle names updated.
+  - Better display of lifetimes and widths.
+  - More tests.
 * Demo notebook added, with a launcher for Binder.
 * Extra tests for particle searches.